+++ tests/nix_tests/network.py
@@ -10,7 +10,14 @@ from pystarport import ports
 from web3.middleware import geth_poa_middleware
 
 from .cosmoscli import CosmosCLI
-from .utils import http_wait_for_block, memiavl_config, supervisorctl, wait_for_port
+from .utils import (
+    EVMOS_6DEC_CHAIN_ID,
+    evm6dec_config,
+    http_wait_for_block,
+    memiavl_config,
+    supervisorctl,
+    wait_for_port,
+)
 
 DEFAULT_CHAIN_BINARY = "evmosd"
 
@@ -127,6 +134,18 @@ def setup_evmos(path, base_port, long_timeout_commit=False):
     yield from setup_custom_evmos(path, base_port, cfg)
 
 
+def setup_evmos_6dec(path, base_port, long_timeout_commit=False):
+    """
+    setup_evmos_6dec returns an Evmos chain with
+    an EVM with 6 decimals and a "0.1" base fee.
+    """
+    config = evm6dec_config(
+        path, "default" if long_timeout_commit is False else "long_timeout_commit"
+    )
+    cfg = Path(__file__).parent / config
+    yield from setup_custom_evmos(path, base_port, cfg, chain_id=EVMOS_6DEC_CHAIN_ID)
+
+
 # for memiavl need to create the data/snapshots dir
 # for the nodes
 def create_snapshots_dir(
@@ -186,7 +205,13 @@ def setup_geth(path, base_port):
 
 
 def setup_custom_evmos(
-    path, base_port, config, post_init=None, chain_binary=None, wait_port=True
+    path,
+    base_port,
+    config,
+    post_init=None,
+    chain_binary=None,
+    wait_port=True,
+    chain_id="evmos_9002-1",
 ):
     cmd = [
         "pystarport",
@@ -214,9 +239,7 @@ def setup_custom_evmos(
             # wait for blocks
             # cause with sdkv0.50 the port starts faster
             http_wait_for_block(ports.rpc_port(base_port), 2)
-        yield Evmos(
-            path / "evmos_9002-1", chain_binary=chain_binary or DEFAULT_CHAIN_BINARY
-        )
+        yield Evmos(path / chain_id, chain_binary=chain_binary or DEFAULT_CHAIN_BINARY)
     finally:
         os.killpg(os.getpgid(proc.pid), signal.SIGTERM)
         proc.wait()