+++ tests/nix_tests/test_cli.py
@@ -1,4 +1,4 @@
-from .utils import get_current_height, supervisorctl, wait_for_block
+from .utils import get_current_height, get_scaling_factor, supervisorctl, wait_for_block
 
 
 def test_block_cmd(evmos_cluster):
@@ -18,7 +18,11 @@ def test_block_cmd(evmos_cluster):
     wait_for_block(node1, last_block)
 
     # stop node1
-    supervisorctl(evmos_cluster.base_dir / "../tasks.ini", "stop", "evmos_9002-1-node1")
+    supervisorctl(
+        evmos_cluster.base_dir / "../tasks.ini",
+        "stop",
+        f"{evmos_cluster.cosmos_cli().chain_id}-node1",
+    )
 
     # use 'block' CLI cmd in node1
     test_cases = [
@@ -59,7 +63,9 @@ def test_block_cmd(evmos_cluster):
 
     # start node1 again
     supervisorctl(
-        evmos_cluster.base_dir / "../tasks.ini", "start", "evmos_9002-1-node1"
+        evmos_cluster.base_dir / "../tasks.ini",
+        "start",
+        f"{evmos_cluster.cosmos_cli().chain_id}-node1",
     )
     # check if chain continues alright
     wait_for_block(node1, last_block + 3)
@@ -73,11 +79,16 @@ def test_tx_flags(evmos_cluster):
     node = evmos_cluster.cosmos_cli(0)
     current_height = get_current_height(node)
     wait_for_block(node, current_height + 1)
+    fee_denom = node.evm_denom()
+    scale_factor = get_scaling_factor(node)
 
     test_cases = [
         {
             "name": "fail - invalid flags combination (gas-prices & fees)",
-            "flags": {"fees": "5000000000aevmos", "gas_prices": "50000aevmos"},
+            "flags": {
+                "fees": f"{int(5000000000 / scale_factor)}{fee_denom}",
+                "gas_prices": f"{50000/ scale_factor}{fee_denom}",
+            },
             "exp_err": True,
             "err_msg": "cannot provide both fees and gas prices",
         },
@@ -89,20 +100,28 @@ def test_tx_flags(evmos_cluster):
         },
         {
             "name": "fail - insufficient fees",
-            "flags": {"fees": "10aevmos", "gas": 50000, "gas_prices": None},
+            "flags": {
+                "fees": f"{int(10/scale_factor)}{fee_denom}",
+                "gas": 50000,
+                "gas_prices": None,
+            },
             "exp_err": True,
             "err_msg": "insufficient fee",
         },
         {
             "name": "fail - insufficient gas",
-            "flags": {"fees": "500000000000aevmos", "gas": 1, "gas_prices": None},
+            "flags": {
+                "fees": f"{int(500000000000/scale_factor)}{fee_denom}",
+                "gas": 1,
+                "gas_prices": None,
+            },
             "exp_err": True,
             "err_msg": "out of gas",
         },
         {
             "name": "success - defined fees & gas",
             "flags": {
-                "fees": "10000000000000000000aevmos",
+                "fees": f"{int(10000000000000000000/scale_factor)}{fee_denom}",
                 "gas": 1500000,
                 "gas_prices": None,
             },
@@ -111,7 +130,10 @@ def test_tx_flags(evmos_cluster):
         },
         {
             "name": "success - using gas & gas-prices",
-            "flags": {"gas_prices": "100000000000aevmos", "gas": 1500000},
+            "flags": {
+                "gas_prices": f"{100000000000/scale_factor}{fee_denom}",
+                "gas": 1500000,
+            },
             "exp_err": False,
             "err_msg": None,
         },
@@ -119,7 +141,7 @@ def test_tx_flags(evmos_cluster):
             "name": "success - using gas 'auto' and specific fees",
             "flags": {
                 "gas": "auto",
-                "fees": "10000000000000000000aevmos",
+                "fees": f"{int(10000000000000000000/scale_factor)}{fee_denom}",
                 "gas_prices": None,
             },
             "exp_err": False,
@@ -132,7 +154,7 @@ def test_tx_flags(evmos_cluster):
             res = node.transfer(
                 "signer1",
                 "evmos10jmp6sgh4cc6zt3e8gw05wavvejgr5pwjnpcky",
-                "100000000000000aevmos",
+                f"{int(100000000000000/scale_factor)}{fee_denom}",
                 False,
                 **tc["flags"],
             )